test: harden live region + playwright capture test
